commit to next level,add user

+++ application/models/M_master_data.php
@@ -4,7 +4,7 @@ class M_master_data extends CI_Model
 
     function get_all_mhs()
     {
-        $this->db->select('*');
+        $this->db->select('a.*,b.nim as nim, b.program_studi as program_studi_id, c.program_studi as program_studi, d.id as program_id,d.nama as nama');
         $this->db->from('mahasiswa a');
         $this->db->join('data_mahasiswa_kampus AS b', 'b.id_mahasiswa=a.id', 'left outer');
         $this->db->join('program_studi AS c', 'b.program_studi=c.id', 'left outer');
@@ -12,12 +12,29 @@ class M_master_data extends CI_Model
         $query = $this->db->get();
         return $query->result_array();
     }
+    function get_all_karyawan()
+    {
+        $this->db->select('a.*,b.nip as nip,b.program_studi as prodi_id ,c.id as id_prodi,d.id as program_id,e.nama as nama');
+        $this->db->from('karyawan a');
+        $this->db->join('data_karyawan_kampus as b', 'b.id_karyawan=a.id', 'left outer');
+        $this->db->join('program_studi as c', 'b.program_studi=c.id', 'left outer');
+        $this->db->join('program as d', 'c.id_program=d.id', 'left outer');
+        $this->db->join('jabatan as e', 'e.id=a.id_jabatan', 'left outer');
+        $query = $this->db->get();
+        return $query->result_array();
+        // $sql = "SELECT a.*, b.id as id_jabatan,b.nama as nama from karyawan a, jabatan b where a.id_jabatan=b.id";
+        // return $this->db->query($sql)->result_array();
+    }
     function get_prodi_by_id($id_program)
     {
     }
     function get_all_program()
     {
         return $this->db->get('program')->result_array();
+    }
+    function get_all_permohonan()
+    {
+        return $this->db->get('jenisPermohonan')->result_array();
     }
     function get_prodi_id($id)
     {
@@ -29,16 +46,7 @@ class M_master_data extends CI_Model
         $query = $this->db->select('*')->from('mahasiswa')->get();
         return $query->result_array();
     }
-    function get_all_karyawan()
-    {
-        // $this->db->select('*');
-        // $this->db->from('karyawan');
-        // $this->db->join('jabatan AS b', 'karyawan.id_jabatan = b.id', 'left outer');
-        // $query = $this->db->get('karyawan');
-        // return $query->result_array();
-        $sql = "SELECT a.*, b.id as id_jabatan,b.nama as nama from karyawan a, jabatan b where a.id_jabatan=b.id";
-        return $this->db->query($sql)->result_array();
-    }
+
     function get_all_jbtn()
     {
         return $this->db->get('jabatan')->result_array();
@@ -47,46 +55,61 @@ class M_master_data extends CI_Model
     function save_karyawan($data)
     {
         $this->db->insert('karyawan', $data);
-        return true;
+        $insert_id = $this->db->insert_id();
+        return $insert_id;
     }
+
     function save_karyawan1($data)
     {
         $this->db->insert('karyawan', $data);
-        return true;
+        $insert_id = $this->db->insert_id();
+        return $insert_id;
     }
+
     function save_mahasiswa($data1)
     {
         $data_mhs = $this->db->insert('mahasiswa', $data1);
         $insert_id = $this->db->insert_id();
         return $insert_id;
-        // if ($data_mhs->affected_rows() > 0) {
-        //     $data2['id_mahasiswa'] = $insert_id;
-        //     array_push($id_mahasiswa, $data2);
-        //     $this->db->insert('data_mahasiswa_kampus', $id_mahasiswa);        
-        //     return true;
-        // } else {
-        //     return false;
-        // }
     }
     function save_mahasiswa1($data1)
     {
         $data_mhs = $this->db->insert('mahasiswa', $data1);
         $insert_id = $this->db->insert_id();
         return $insert_id;
-        // if ($data_mhs) {
-        //     $data2['id_mahasiswa'] = $insert_id;
-        //     array_push($id_mahasiswa, $data2);
-        //     $this->db->insert('data_mahasiswa_kampus', $id_mahasiswa);
-        //     return true;
-        // } else {
-        //     return false;
-        // }
     }
     function save_data_kampus($data2)
     {
         $insert_id = $this->db->insert('data_mahasiswa_kampus', $data2);
         return true;
     }
+    function save_data_karyawan($data2)
+    {
+        $insert_id = $this->db->insert('data_karyawan_kampus', $data2);
+        return true;
+    }
+
+    function update_data_kampus($data2, $id)
+    {
+        $this->db->where('id_mahasiswa', $id);
+        $this->db->update('data_mahasiswa_kampus', $data2);
+        if ($this->db->affected_rows()) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+    function update_data_kampus1($data2, $id)
+    {
+        $this->db->where('id_mahasiswa', $id);
+        $this->db->update('data_mahasiswa_kampus', $data2);
+        if ($this->db->affected_rows()) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
     public function update_mahasiswa($data, $id)
     {
         $this->db->where('id', $id);
@@ -128,14 +151,31 @@ class M_master_data extends CI_Model
         }
     }
 
-    function delete_mahasiswa($id)
+    function delete_data_mahasiswa($id)
     {
-        $this->db->where('id', $id);
-        $this->db->delete('mahasiswa');
+        $this->db->where('id_mahasiswa', $id);
+        $a = $this->db->delete('data_mahasiswa_kampus');
+
+        if ($a) {
+            $this->db->where('id', $id);
+            $this->db->delete('mahasiswa');
+            return true;
+        } else {
+            return false;
+        }
     }
+
     function delete_karyawan($id)
     {
-        $this->db->where('id', $id);
-        $this->db->delete('karyawan');
+        $this->db->where('id_karyawan', $id);
+        $a = $this->db->delete('data_karyawan_kampus');
+
+        if ($a) {
+            $this->db->where('id', $id);
+            $this->db->delete('karyawan');
+            return true;
+        } else {
+            return false;
+        }
     }
 }
